+++ src/main/java/com/sandymandy/pleasurecraft/registries/GirlRegistry.java
@@ -1,4 +1,4 @@
-package com.sandymandy.pleasurecraft.util;
+package com.sandymandy.pleasurecraft.registries;
 
 import com.sandymandy.pleasurecraft.entity.PleasureCraftEntities;
 import com.sandymandy.pleasurecraft.entity.girls.MikaEntity;